feat: introduce useAppEvent hook and deprecate useFrame
- Added a new `useAppEvent` hook for subscribing to PlayCanvas application events with flexible parameter signatures.
- Deprecated the `useFrame` hook, recommending the use of `useAppEvent('update', callback)` instead.
- Updated exports to reflect the changes and removed the old `useFrame` implementation.
