+++ .changeset/vast-roses-wait.md
@@ -1,0 +1,5 @@
+---
+"@playcanvas/react": minor
+---
+
+Deprecated useFrame and providing a new useAppEvent api
