+++ packages/lib/src/hooks/index.ts
@@ -6,5 +6,5 @@ export { useApp, AppContext } from './use-app.tsx';
 export { useParent, ParentContext } from './use-parent.tsx';
 export { useMaterial } from './use-material.tsx'
 export { useAsset, useSplat, useTexture, useEnvAtlas, useModel, useFont } from './use-asset.ts';
-export { useFrame } from './use-frame.tsx';
+export { useFrame, useAppEvent } from './use-app-event.ts';
 export type { AssetResult } from './use-asset.ts';
