+++ packages/lib/src/hooks/use-frame.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { renderHook } from '@testing-library/react';
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { useFrame } from './use-frame.tsx';
+import { useFrame } from './use-app-event.ts';
 import { Application } from '../Application.tsx';
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 
 describe('useFrame', () => {
